backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
yt-dlp
python-dotenv==1.0.1
bcrypt==4.2.0
Werkzeug==3.0.4
google-auth==2.29.0
requests==2.31.0
gunicorn==22.0.0
waitress==2.1.2

